rename NodeFilterStore to NodeFilterLocalstore
